requirements.txt
Flask==3.1.2
supabase==2.3.4
bcrypt==4.1.2
python-dotenv==1.0.0
gunicorn==21.2.0